+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## 0.12.0 [unreleased] 
 
+### Features
+
+1. [#123](https://github.com/InfluxCommunity/influxdb3-python/pull/123): Introduces `query_async()` method. From this release the client now has a `query_async()` method that takes advantage of asyncio's event loop to run query calls in their own executor.
+
+For example:
+```python
+    table = await client.query_async(query)
+```
+
 ### Bug Fixes
 
 1. [#121](https://github.com/InfluxCommunity/influxdb3-python/pull/121): Fix use of arguments `verify_ssl` and `ssl_ca_cert` in `QueryApi`.